@@ -23,8 +23,8 @@
   "repository": "https://github.com/esp/esp-js-react",
   "typings": "./esp-js-react.d.ts",
   "peerDependencies": {
-    "esp-js": "^1.1.3",
-    "react": "^0.14.0 || ^15.0.0"
+    "esp-js": "^1.4.0",
+    "react": "^16.0.0"
   },
   "devDependencies": {
     "babel-core": "^6.4.5",
@@ -38,15 +38,19 @@
     "copy-webpack-plugin": "^1.1.1",
     "eslint": "^1.10.3",
     "eslint-loader": "^1.2.1",
+    "esp-js": "^1.4.0",
     "inject-loader": "^2.0.1",
     "jasmine-core": "^2.3.4",
     "karma": "^0.13.19",
     "karma-chrome-launcher": "^0.2.2",
     "karma-jasmine": "^0.3.7",
-    "karma-spec-reporter": "0.0.23",
+    "karma-spec-reporter": "0.0.31",
     "karma-webpack": "^1.7.0",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
     "source-map-loader": "^0.1.5",
     "webpack": "^1.12.12",
     "webpack-dev-server": "^1.14.1"
-  }
+  },
+  "dependencies": {}
 }
